[CLOUD-2982] Convert EAP CD Arquillian tests to EAP 7.2
