+++ tests/arq/pom.xml
@@ -11,11 +11,11 @@
     </parent>
 
     <groupId>org.jboss.ce.testsuite</groupId>
-    <artifactId>eap-cd-arq-testsuite</artifactId>
+    <artifactId>eap72-arq-testsuite</artifactId>
     <packaging>jar</packaging>
-    <name>EAP CD testsuite</name>
+    <name>EAP 7.2 testsuite</name>
     <version>1.0-SNAPSHOT</version>
-    <description>Aruqllian EAP CD testsuite</description>
+    <description>Arquillian EAP 7.2 testsuite</description>
 
     <properties>
         <version.arquillian>1.4.0.Final</version.arquillian>
@@ -25,9 +25,10 @@
         <version.org.jboss.as.arquillian.container>7.2.0.Final</version.org.jboss.as.arquillian.container>
         <version.ce-arq>1.3.0.Final</version.ce-arq>
 
-        <image>registry.access.redhat.com/jboss-eap-7-tech-preview/eap-cd-openshift</image>
-        <image.stream.version>14</image.stream.version>
-        <image.stream.name>eap-cd-openshift</image.stream.name>
+        <image>registry.access.redhat.com/jboss-eap-7/eap72-openshift</image>
+        <image.stream.version>1.0</image.stream.version>
+        <image.stream.name>eap72-openshift</image.stream.name>
+        <template.prefix>eap72</template.prefix>
         <forkCount>5</forkCount>
     </properties>
 
@@ -151,7 +152,8 @@
                         <image.stream.image>${image}</image.stream.image>
                         <image.stream.name>${image.stream.name}</image.stream.name>
                         <image.stream.version>${image.stream.version}</image.stream.version>
-                        <template.uri>https://raw.githubusercontent.com/jboss-container-images/jboss-eap-7-openshift-image/eap-cd-dev/templates</template.uri>
+                        <template.prefix>${template.prefix}</template.prefix>
+                        <template.uri>https://raw.githubusercontent.com/jboss-container-images/jboss-eap-7-openshift-image/eap72-dev/templates</template.uri>
                         <secrets.repository>jboss-openshift</secrets.repository>
                         <secrets.branch>master</secrets.branch>
                         <!-- URL to the OpenShift image streams that are officially added to OCP. These image streams are added to every test project run on OpenShift -->
